feat: implement DormPass and Payment service implementations with CRUD operations

+++ src/main/java/com/dormhub/model/Payment.java
@@ -30,7 +30,7 @@ public class Payment {
         return amount;
     }
 
-    public void setAmount(int amount) {
+    public void setAmount(double amount) {
         this.amount = amount;
     }
